+++ HISTORY.md
@@ -13,6 +13,7 @@ response attributes instead of just a code. To access the authorization code, ge
 * `OAuth2Client.authorization_code()` now accepts an `AuthorizationResponse` as parameter, and will
 use it to include all necessary parameters for the Authorization Code Grant.
 * removed `OAuth2Client.authorization_code_pkce()`
+* Added `ClientSecretJwt` and `PrivateKeyJwt` aliases to `ClientSecretJWT` and `PrivateKeyJWT`, for consistency with `jwskate`.
 ## <= 0.18
 
 * Draft versions